+++ test/test_importer.py
@@ -53,15 +53,21 @@ class AutotagStub(object):
     def install(self):
         self.mb_match_album = autotag.mb.match_album
         self.mb_match_track = autotag.mb.match_track
+        self.mb_album_for_id = autotag.mb.album_for_id
+        self.mb_track_for_id = autotag.mb.track_for_id
 
         autotag.mb.match_album = self.match_album
         autotag.mb.match_track = self.match_track
+        autotag.mb.album_for_id = self.album_for_id
+        autotag.mb.track_for_id = self.track_for_id
 
         return self
 
     def restore(self):
         autotag.mb.match_album = self.mb_match_album
         autotag.mb.match_track = self.mb_match_album
+        autotag.mb.album_for_id = self.mb_album_for_id
+        autotag.mb.track_for_id = self.mb_track_for_id
 
     def match_album(self, albumartist, album, tracks):
         if self.matching == self.IDENT:
@@ -86,6 +92,12 @@ class AutotagStub(object):
             artist_id=u'artistid',
             length=1
         )
+
+    def album_for_id(self, mbid):
+        return None
+
+    def track_for_id(self, mbid):
+        return None
 
     def _make_track_match(self, artist, album, number):
         return TrackInfo(
@@ -970,6 +982,23 @@ class InferAlbumDataTest(_common.TestCase):
         self.assertFalse(self.items[0].comp)
 
 
+def test_album_info():
+    """Create an AlbumInfo object for testing.
+    """
+    track_info = TrackInfo(
+        title=u'new title',
+        track_id=u'trackid',
+    )
+    album_info = AlbumInfo(
+        artist=u'artist',
+        album=u'album',
+        tracks=[track_info],
+        album_id=u'albumid',
+        artist_id=u'artistid',
+    )
+    return album_info
+
+
 class ImportDuplicateAlbumTest(unittest.TestCase, TestHelper):
 
     def setUp(self):
@@ -981,18 +1010,7 @@ class ImportDuplicateAlbumTest(unittest.TestCase, TestHelper):
         # Create duplicate through autotagger
         self.match_album_patcher = patch('beets.autotag.mb.match_album')
         self.match_album = self.match_album_patcher.start()
-        track_info = TrackInfo(
-            title=u'new title',
-            track_id=u'trackid',
-        )
-        album_info = AlbumInfo(
-            artist=u'artist',
-            album=u'album',
-            tracks=[track_info],
-            album_id=u'albumid',
-            artist_id=u'artistid',
-        )
-        self.match_album.return_value = iter([album_info])
+        self.match_album.return_value = iter([test_album_info()])
 
         # Create import session
         self.importer = self.create_importer()
@@ -1370,6 +1388,89 @@ class MultiDiscAlbumsInDirTest(_common.TestCase):
             os.remove(path)
         albums = list(albums_in_dir(self.base))
         self.assertEquals(len(albums), 0)
+
+
+class ReimportTest(unittest.TestCase, ImportHelper):
+    """Test "re-imports", in which the autotagging machinery is used for
+    music that's already in the library.
+
+    This works by importing new database entries for the same files and
+    replacing the old data with the new data. We also copy over flexible
+    attributes and the added date.
+    """
+
+    def setUp(self):
+        self.setup_beets()
+
+        # The existing album.
+        album = self.add_album_fixture()
+        album.added = 4242.0
+        album.foo = u'bar'  # Some flexible attribute.
+        album.store()
+        item = album.items().get()
+        item.baz = u'qux'
+        item.added = 4747.0
+        item.store()
+
+        # Set up an import pipeline with a "good" match.
+        self.matcher = AutotagStub().install()
+        self.matcher.matching = AutotagStub.GOOD
+
+    def tearDown(self):
+        self.teardown_beets()
+        self.matcher.restore()
+
+    def _setup_session(self, singletons=False):
+        self._setup_import_session(self._album().path, singletons=singletons)
+        self.importer.add_choice(importer.action.APPLY)
+
+    def _album(self):
+        return self.lib.albums().get()
+
+    def _item(self):
+        return self.lib.items().get()
+
+    def test_reimported_album_gets_new_metadata(self):
+        self._setup_session()
+        self.assertEqual(self._album().album, u'\xe4lbum')
+        self.importer.run()
+        self.assertEqual(self._album().album, u'the album')
+
+    def test_reimported_album_preserves_flexattr(self):
+        self._setup_session()
+        self.importer.run()
+        self.assertEqual(self._album().foo, u'bar')
+
+    def test_reimported_album_preserves_added(self):
+        self._setup_session()
+        self.importer.run()
+        self.assertEqual(self._album().added, 4242.0)
+
+    def test_reimported_album_preserves_item_flexattr(self):
+        self._setup_session()
+        self.importer.run()
+        self.assertEqual(self._item().baz, u'qux')
+
+    def test_reimported_album_preserves_item_added(self):
+        self._setup_session()
+        self.importer.run()
+        self.assertEqual(self._item().added, 4747.0)
+
+    def test_reimported_item_gets_new_metadata(self):
+        self._setup_session(True)
+        self.assertEqual(self._item().title, u't\xeftle 0')
+        self.importer.run()
+        self.assertEqual(self._item().title, u'full')
+
+    def test_reimported_item_preserves_flexattr(self):
+        self._setup_session(True)
+        self.importer.run()
+        self.assertEqual(self._item().baz, u'qux')
+
+    def test_reimported_item_preserves_added(self):
+        self._setup_session(True)
+        self.importer.run()
+        self.assertEqual(self._item().added, 4747.0)
 
 
 def suite():